Add MSTest wasm acceptance test and install wasm-tools workload in CI
Split the wasi-wasm acceptance coverage per product: the raw-platform tests stay in Microsoft.Testing.Platform.Acceptance.IntegrationTests and the MSTest-adapter tests move to MSTest.Acceptance.IntegrationTests. Shared wasmtime/publish/ICU plumbing is extracted into a WasmRuntime helper in Microsoft.Testing.TestInfrastructure to avoid duplication across the two projects. Also add the wasm-tools-net10 workload to eng/restore-toolset so 'dotnet publish -r wasi-wasm' works in CI (execution still self-skips unless wasmtime is on PATH).
